+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Patternbot adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 
 - The output file can now be customized so it isn’t always named `pattern-library.html`. It can be changed to `index.html` for easier library loading.
+- Code blocks are now collapsed by default and can be expanded with a button.
 
 ### Fixed